updated practical test 1-6
